[PATCH] ppc32: fix ebony.c warnings
This patch removes annoying warnings in ebony.c.  Fix is similar to one I
made in ocotea.c before.

+++ arch/ppc/platforms/4xx/ebony.c
@@ -128,7 +128,7 @@ ebony_map_irq(struct pci_dev *dev, unsigned char idsel, unsigned char pin)
 }
 
 #define PCIX_WRITEL(value, offset) \
-	(writel(value, (u32)pcix_reg_base+offset))
+	(writel(value, pcix_reg_base + offset))
 
 /*
  * FIXME: This is only here to "make it work".  This will move
